% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{medium_auth}
\alias{medium_auth}
\alias{medium_set_auth}
\title{Medium Authentication}
\usage{
medium_set_auth()
}
\description{
\code{medium_set_auth()} interactivery asks for the token and store it as an environmental variable \code{MEDIUM_API_TOKEN}.
It is recommended to set this variable in \code{.Renviron}.
}
\seealso{
\url{https://github.com/Medium/medium-api-docs#22-self-issued-access-tokens}
}
