% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addin.R
\name{medium_create_post_from_Rmd}
\alias{medium_create_post_from_Rmd}
\title{Publish R Markdown Document to 'Medium'}
\usage{
medium_create_post_from_Rmd(Rmd_file = NULL)
}
\arguments{
\item{Rmd_file}{path to a .Rmd file. If \code{NULL}, use the activedocument.}
}
\description{
Knit and post a given R Markdown file to 'Medium'.
}
