% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image.R
\name{medium_image}
\alias{medium_image}
\alias{medium_upload_image}
\title{Medium Images API}
\usage{
medium_upload_image(image)
}
\arguments{
\item{image}{A path to a image file or \code{form_file} object.}
}
\description{
Uploading an image.
}
\examples{
\dontrun{
medium_upload_image("path/to/image.png")
medium_upload_image(httr::upload_file("path/to/image.png"))
}
}
\seealso{
\url{https://github.com/Medium/medium-api-docs#34-images}
}
