% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post.R
\name{medium_post}
\alias{medium_post}
\alias{medium_create_post}
\title{Medium Posts API}
\usage{
medium_create_post(title, contentFormat = c("markdown", "html"),
  content = "", tags = c("R", "rstats"), canonicalUrl = NULL,
  publishStatus = c("draft", "public", "unlisted"),
  license = c("all-rights-reserved", "cc-40-by", "cc-40-by-sa", "cc-40-by-nd",
  "cc-40-by-nc", "cc-40-by-nc-nd", "cc-40-by-nc-sa", "cc-40-zero",
  "public-domain"), notifyFollowers = FALSE, publicationId = NULL)
}
\arguments{
\item{title}{The title of the post. Note that this title is used for SEO and when rendering the post as a listing,
but will not appear in the actual post. For that, the title must be specified in the \code{content} field as well.
Titles longer than 100 characters will be ignored. In that case, a title will be synthesized from the first
content in the post when it is published.}

\item{contentFormat}{The format of the \code{content} field. There are two valid values, \code{"html"}, and \code{"markdown"}.}

\item{content}{The body of the post, in a valid, semantic, HTML fragment, or Markdown. Further markups may be supported
in the future. For a full list of accepted HTML tags, see \href{https://medium.com/@katie/a4367010924e}{here}.
If you want your title to appear on the post page, you must also include it as part of the post content.}

\item{tags}{Tags to classify the post. Only the first three will be used. Tags longer than 25 characters will be ignored.}

\item{canonicalUrl}{The original home of this content, if it was originally published elsewhere.}

\item{publishStatus}{The status of the post. Valid values are \code{"public"}, \code{"draft"}, or \code{"unlisted"}. The default is \code{"draft"}.}

\item{license}{The license of the post. Valid values are \code{"all-rights-reserved"}, \code{"cc-40-by"}, \code{"cc-40-by-sa"}, \code{"cc-40-by-nd"},
\code{"cc-40-by-nc"}, \code{"cc-40-by-nc-nd"}, \code{"cc-40-by-nc-sa"}, \code{"cc-40-zero"}, \code{"public-domain"}.
The default is \code{"all-rights-reserved"}.}

\item{notifyFollowers}{Whether to notifyFollowers that the user has published.}

\item{publicationId}{Publication ID.}
}
\description{
Creates a post on the authenticated user's profile or a given publication.
}
\examples{
\dontrun{
content <- "
# test

1. test1
2. test2

```r
this <- is(test)
```
"
medium_create_post("test", content = content)
}

}
\seealso{
\url{https://github.com/Medium/medium-api-docs#33-posts}
}
