% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/publication.R
\name{medium_publication}
\alias{medium_publication}
\alias{medium_get_publications}
\alias{medium_get_contributors}
\title{Medium Publications API}
\usage{
medium_get_publications(userId = Sys.getenv("MEDIUM_USER_ID"))

medium_get_contributors(publicationId)
}
\arguments{
\item{userId}{User ID (no need to specify this usually).}

\item{publicationId}{Publication ID.}
}
\value{
Publication:
\itemize{
\item \code{id}: A unique identifier for the publication.
\item \code{name}: The publication's name on Medium.
\item \code{description}: Short description of the publication
\item \code{url}: The URL to the publication's homepage
\item \code{imageUrl}: The URL to the publication's image/logo
Contributor:
\item \code{publicationId}: An ID for the publication.
\item \code{userId}: A user ID of the contributor.
\item \code{role}: Role of the user identified by userId in the publication identified by publicationId. 'editor' or 'writer'.
}
}
\description{
\code{medium_get_publications()} returns a full list of publications that the user is related to in some way.
\code{medium_get_contributors()} returns a list of contributors for a given publication.
}
