% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addin.R
\name{medium_upload_Rmd}
\alias{medium_upload_Rmd}
\title{Publish R Markdown Document to 'Medium'}
\usage{
medium_upload_Rmd(input = NULL)
}
\arguments{
\item{input}{path to a .Rmd file. If \code{NULL}, use the activedocument.}
}
\description{
Knit and post a given R Markdown file to 'Medium'.
}
