% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user.R
\name{medium_user}
\alias{medium_user}
\alias{medium_get_current_user}
\title{Medium Users API}
\usage{
medium_get_current_user()
}
\value{
\itemize{
\item \code{id}: A unique identifier for the user.
\item \code{username}: The user’s username on Medium.
\item \code{name}: The user’s name on Medium.
\item \code{url}: The URL to the user’s profile on Medium
\item \code{imageUrl}: The URL to the user’s avatar on Medium
}
}
\description{
Returns details of the user who has granted permission to the application.
}
\seealso{
\url{https://github.com/Medium/medium-api-docs#getting-the-authenticated-users-details}
}
