% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediumr-package.R
\docType{package}
\name{mediumr-package}
\alias{mediumr-package}
\alias{_PACKAGE}
\alias{mediumr}
\title{R Interface to 'Medium' API}
\description{
Provides an interface to 'Medium' API.
}
\author{
\strong{Maintainer}: Hiroaki Yutani \email{yutani.ini@gmail.com}

}
